[#86] 활동 삭제 및 수정 기능 수정

+++ src/main/java/com/hobak/happinessql/domain/activity/application/ActivityDeleteService.java
@@ -4,6 +4,7 @@ package com.hobak.happinessql.domain.activity.application;
 import com.hobak.happinessql.domain.activity.domain.Activity;
 import com.hobak.happinessql.domain.activity.exception.ActivityNotFoundException;
 import com.hobak.happinessql.domain.activity.repository.ActivityRepository;
+import com.hobak.happinessql.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,9 +14,13 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ActivityDeleteService {
     private final ActivityRepository activityRepository;
-    public void deleteActivity(Long activityId){
+    public Long deleteActivity(Long activityId, User user){
         Activity activity = activityRepository.findById(activityId)
                 .orElseThrow(()->new ActivityNotFoundException("Activity with id " + activityId));
-        activityRepository.delete(activity);
+        if(activity.getCategory().getUserId().equals(user.getUserId())) {
+            activityRepository.delete(activity);
+            return activity.getActivityId();
+        }
+        return null;
     }
 }
